+++ .github/workflows/lsst-tests.yml
@@ -34,8 +34,8 @@ jobs:
           mamba install --quiet --yes --file dev-requirements.txt
           mamba install -q -y stackvana=0 lsstdesc-weaklensingdeblending
 
-          pip install --no-deps git+https://github.com/LSSTDESC/descwl-shear-sims.git
-          pip install --no-deps git+https://github.com/LSSTDESC/descwl_coadd.git
+          pip install --no-deps git+https://github.com/LSSTDESC/descwl-shear-sims.git@mockpsf
+          pip install --no-deps git+https://github.com/LSSTDESC/descwl_coadd.git@exp-inputs
 
           python -m pip install -e .
