+++ shear_meas_test/test_shear_meas.py
@@ -256,7 +256,7 @@ def test_shear_meas(snr, ngrid, ntrial):
 
     tm0 = time.time()
 
-    print("\n")
+    print("")
 
     pres = []
     mres = []
